0.0.7

+++ __init__.py
@@ -4,7 +4,7 @@ import threading
 import rich.progress
 import rich.text
 import os
-from typing import *
+from typing import Callable,Any,List
 import tempfile
 import random
 import time
@@ -51,14 +51,16 @@ class _Part:
     A download task. It mean a part of the file we are downloading.
     It include the range of this task, the response, the statue, etc.
     """
-    def __init__(self,start_:int,to:int,fileName:str|None=None,stream:None|requests.Response=None) -> None:
+    def __init__(self,start_:int,to:int,num:int=0,fileName:str="",stream:None|requests.Response=None) -> None:
         """
         New a Part object
         :param start_: the start position of the part
         :param to: the end position of the part
+        :param num: the num of the part
         :param fileName: the tempfile name of the part
         :param stream: the response of the part
         """
+        self.num=num
         self.start=start_
         self.to=to
         self.fileName=fileName
@@ -89,7 +91,7 @@ class _Part:
             return  self.start<other.start
         return self.to<other.to
 class AutoDownload:
-    def __init__(self,url:str,file:str,chunkSize:int=1024,maxRetry:int=5,continueDownloadTest:bool=False,startSize:int=0,openType:str="wb",
+    def __init__(self,url:str,file:str,chunkSize:int=1024,maxRetry:int=5,maxThreadRetry:int=-1,timeout:int|None=30,continueDownloadTest:bool=False,startSize:int=0,openType:str="wb",
                  error:bool=True,log:bool=True,showProgressBar:bool=True,transient:bool=False,
                  threaded:bool=False,threadNum:int=0,maxThreadNum:int=10,desiredCompletionTime:int=30,
                  callbackFunction:None|Callable[[bool], Any]=None,deamon:bool=False,header:dict={})->None:
@@ -98,8 +100,9 @@ class AutoDownload:
         :param url: url to download
         :param file: file name
         :param chunkSize: chunk size
-        :param maxRetry: max retry times
-        :param continueDownloadTest: Whether to detect power interruption
+        :param maxRetry: max retry times for the first connection. If it's less than 0, it means infinity
+        :param maxThreadRetry: max retry times for the other connection. If it's less than 0, it means infinity
+        :param continueDownloadTest: Whether to detect power interruption.
         :param startSize: where to start downloading. If continueDownloadTest is True, it will be covered
         :param openType: file open type. If continueDownloadTest is True, it will be covered
         :param error: whether to raise error
@@ -111,7 +114,7 @@ class AutoDownload:
         :param threadNum: num of thread. If threadNum < 1, we'll automatically calculates the threadNum based on the total size and the speed of the first thread
         :param maxThreadNum: max num of thread. It will be overlooked if threadNum >= 1
         :param desiredCompletionTime: time in seconds. It is our reference value for calculating the threadNum
-        :param deamon: whether to run in deamon mode
+        :param deamon: whether to run in deamon mode.
         :param header: request header
         """
         self.url = url
@@ -134,6 +137,9 @@ class AutoDownload:
         self._waitList:List[int]=[]
         self.logger=logging.getLogger("Download")
         self.openType=openType
+        self.timeout=timeout
+        self.maxThreadRetry=maxThreadRetry
+        self.fileSize:int=-1
         self.progress=rich.progress.Progress(
             rich.progress.TextColumn("[progress.description]{task.description}"),
             rich.progress.BarColumn(),
@@ -149,6 +155,13 @@ class AutoDownload:
             disable=not showProgressBar
         )
         self.tempFileDir=os.path.join(tempfile.gettempdir(),self.url.split("/")[-1].split("?")[0]+str(random.random()))
+        self.now=0
+        self.speed=0
+        self.speeds=""
+        self.statue="init"
+        self.historyTime=1
+        self.historyNum=0
+        self.fail=False
     def _logShower(self,msg:str,level=logging.INFO):
         """
         Show log message through with self.logger. And if self.log is False, it will do nothing.
@@ -158,12 +171,13 @@ class AutoDownload:
         if not self.log:
             return
         self.logger.log(level=level,msg=msg)
-    def _errorShower(self,err:Exception):
+    def _errorShower(self,err:BaseException):
         """
         Show error message and raise it.
         :param err: the error
         """
         self._logShower("%s:%s"%(err.__class__.__name__,str(err)),logging.ERROR)
+        self.fail=True
         if self.error:
             raise err
     def start(self)->bool|None:
@@ -188,6 +202,8 @@ class AutoDownload:
         :return: the formatted string 
         """
         units=["B","KB","MB","GB","TB"]
+        if num<0:
+            return "unknown"
         for i in range(len(units)):
             if num/(1024**i)<1024:
                 return "%.2f%s"%((num/(1024**i)),units[i])
@@ -204,6 +220,13 @@ class AutoDownload:
             part.speeds=self.changeUnit(part.speed)+"/s"
             part.histoyTime=t
             part.historyNum=length
+        if t!=self.historyTime and part.num==0:
+            self.speed=self.historyNum
+            self.speeds=self.changeUnit(self.speed)+"/s"
+            self.historyTime=t
+            self.historyNum=length
+        self.historyNum+=length
+        self.now+=length
         part.historyNum+=length
         part.now+=length
     def _updateProgressBar(self):
@@ -213,38 +236,44 @@ class AutoDownload:
         colors=["red","blue","green","gray","purple"]
         statuesColor=["blue","yellow","green","red"]
         with self.progress:
-            total=self.progress.add_task("[yellow]Total",total=self.fileSize,speed="",size="",now="",statue="")
+            self.total=self.progress.add_task("[yellow]Total",total=self.fileSize,start=False,speed="",size="",now="",statue="")
             while True:
                 time.sleep(0.1)
-                sums=0
-                speedsum=0
                 num=0
                 for i in self._partition:
                     num+=1
-                    sums+=i.now
-                    speedsum+=i.speed
                     if i.progress!=None:
                         self.progress.update(i.progress,completed =i.now,total=i.to-i.start,speed=i.speeds,size=self.changeUnit(i.to-i.start),now=self.changeUnit(i.now),statue=f"[{statuesColor[i.statueNum]}]{i.statue}[/{statuesColor[i.statueNum]}]")
                     else:
                         i.progress=self.progress.add_task("[%s]Thread%d"%(random.choice(colors),num),total=i.to-i.start,speed="",size="",now="",statue="init")
-                self.progress.update(total,completed =sums,total=self.fileSize,speed=self.changeUnit(speedsum)+"/s",size=self.changeUnit(self.fileSize),now=self.changeUnit(sums))
+                if self.fileSize>=0:
+                    self.progress.start_task(self.total)
+                self.progress.update(self.total,completed =self.now,total=self.fileSize,speed=self.speeds,size=self.changeUnit(self.fileSize),now=self.changeUnit(self.now),statue=f"[yellow]{self.statue}[/yellow]")
     def _splitThread(self):
+        """
+        Split the first part into multiple threads
+        """
         if self.threadNum<1:
             while self._partition[0].speed==0:
                 time.sleep(1)
             time.sleep(1)
+            if self.maxThreadNum<=0:
+                self._logShower("The maxThreadNum is less than or equal to 0. We try to download it in single thread.",level=logging.WARNING)
+                return
             self.threadNum=min(self.maxThreadNum,int((self.fileSize)//(self._partition[0].speed*self.desiredCompletionTime)))
         if self.threadNum>1:
             if self._partition[0].start+self._partition[0].now+(self.fileSize-self._partition[0].now)//(self.threadNum)>=self._partition[0].to:
                 return
             self._partition.append(self._partition[0].split(self._partition[0].start+self._partition[0].now+(self.fileSize-self._partition[0].now)//(self.threadNum)))
+            self._partition[-1].num=1
             elseSize=self._partition[-1].to-self._partition[-1].start
             for i in range(2,self.threadNum):
                 self._partition.append(self._partition[-1].split(self._partition[-1].start+elseSize//(self.threadNum-1)))
+                self._partition[-1].num=i
                 
             for i in range(1,len(self._partition)):
                 self._partition[i].fileName=os.path.join(self.tempFileDir,f"{i}.tmp")
-                self._threadPool.append(threading.Thread(target=self._download,daemon=self.deamon,args=[i]))
+                self._threadPool.append(threading.Thread(target=self._download,daemon=True,args=[i]))
                 self._waitList.append(i)
                 self._threadPool[-1].start()
     def _wait(self)->bool:
@@ -260,38 +289,40 @@ class AutoDownload:
     def _controller(self)->bool:
         firstHeader=self.header.copy()
         firstHeader["Range"]="bytes=%d-"%(self.startSize)
+        threading.Thread(target=self._updateProgressBar,daemon=True).start()
         for i in range(self.maxRetry):
             try:
-                retsult=requests.get(self.url,headers=firstHeader,stream=True)
+                self.statue="connecting"
+                retsult=requests.get(self.url,headers=firstHeader,stream=True,timeout=self.timeout)
                 if retsult.status_code//100 not in [2,3]:
                     raise ConnectError(self.url)
                 if 'content-length' not in retsult.headers:
                     self._logShower("Can not get the length of the file. try to download normally",level=logging.WARNING)
                     with open(self.file,self.openType) as f:
-                        checkTime=1
-                        now=0
-                        speed=0
-                        speeds="unknown"
-                        with self.progress:
-                            task=self.progress.add_task("[yellow]download",start=False,total=None,speed=self.changeUnit(speed),size="unknown",now="0B",statue="")
-                            for i in retsult.iter_content(chunk_size=self.chunkSize):
-                                now+=len(i)
-                                if checkTime!=int(time.time()):
-                                    checkTime=int(time.time())
-                                    speeds=self.changeUnit(speed)+"/s"
-                                    speed=len(i)
-                                else:
-                                    speed+=len(i)
-                                self.progress.update(task,speed=speeds,statue="[red]downloading",now=self.changeUnit(now))
-                                f.write(i)
-                        self.progress.update(task,speed=speeds,statue="[green]finished",now=self.changeUnit(now))
-                        self.progress.refresh()
+                        self.statue="downloading"
+                        for i in retsult.iter_content(chunk_size=self.chunkSize):
+                            t=int(time.time())
+                            if t!=self.historyTime:
+                                self.speed=self.historyNum
+                                self.speeds=self.changeUnit(self.speed)+"/s"
+                                self.histoyTime=t
+                                self.historyNum=len(i)
+                            self.historyNum+=len(i)
+                            self.now+=len(i)
+                            f.write(i)
+                    
+                    try:
+                        self.progress.update(self.total,speed=self.speeds,statue="[yellow]finished[/yellow]",now=self.changeUnit(self.now))
+                    except:
+                        pass
+                    self.progress.refresh()
                     return True
                 self.fileSize=int(retsult.headers['content-length'])
-                if not self.fileSize:
+                if self.fileSize<=0:
                     raise ZeroDivisionError(self.url)
-                self._partition.append(_Part(self.startSize,self.startSize+int(self.fileSize),os.path.join(self.tempFileDir,"0.tmp"),retsult))
-                self._threadPool.append(threading.Thread(target=self._download,daemon=self.deamon,args=[0]))
+                self.statue="downloading"
+                self._partition.append(_Part(self.startSize,self.startSize+int(self.fileSize),0,os.path.join(self.tempFileDir,"0.tmp"),retsult))
+                self._threadPool.append(threading.Thread(target=self._download,daemon=True,args=[0]))
                 self._waitList.append(0)
                 self._threadPool[0].start()
                 break
@@ -301,27 +332,42 @@ class AutoDownload:
                     self._errorShower(err)
                     return False
         threading.Thread(target=self._splitThread,daemon=True).start()
-        threading.Thread(target=self._updateProgressBar,daemon=True).start()
-        while True:
-            time.sleep(0.5)
-            if not self._waitList:
-                self._logShower("All download finished. Start splicing",level=logging.DEBUG)
-                splicing=self.progress.add_task("[yellow]splicing",total=self.fileSize,speed="",size="",now="",statue="")
-                self._partition.sort()
-                with open(self.file,self.openType) as wf:
-                    for i in self._partition:
-                        with open(i.fileName,"rb") as f:
-                            data=f.read()
-                            data=data[:i.to-i.start]
-                            wf.write(data)
-                            self.progress.update(splicing,advance=len(data))
-                        try:
-                            os.remove(i.fileName)
-                        except:
-                            pass
-                self.progress.update(splicing,statue="[green]finished[/green]")
-                self.progress.refresh()
-                return True
+        try:
+            while True:
+                time.sleep(0.5)
+                if self.fail:
+                    return False
+                if not self._waitList:
+                    self.statue="finished"
+                    self.progress.update(self.total,completed=self.fileSize)
+                    self._logShower("All download finished. Start splicing",level=logging.DEBUG)
+                    splicing=self.progress.add_task("[yellow]splicing",total=self.fileSize,speed="",size="",now="",statue="")
+                    self._partition.sort()
+                    with open(self.file,self.openType) as wf:
+                        for i in self._partition:
+                            now=0
+                            with open(i.fileName,"rb") as f:
+                                while True:
+                                    data=f.read(self.chunkSize)
+                                    if now+len(data)>=i.to-i.start:
+                                        wf.write(data[:i.to-i.start-now])
+                                        self.progress.update(splicing,advance=i.to-i.start-now)
+                                        break
+                                    if not len(data):
+                                        self._errorShower(ValueError("The size of the part is not enough"))
+                                    now+=len(data)
+                                    wf.write(data)
+                                    self.progress.update(splicing,advance=len(data))
+                            try:
+                                os.remove(i.fileName)
+                            except:
+                                pass
+                    self.progress.update(splicing,completed=self.fileSize,statue="[green]finished[/green]")
+                    self.progress.refresh()
+                    return True
+        except BaseException as err:
+            self._errorShower(err)
+            return False
     def _finished(self)->None:
         """
         When a download thread is finished. Find another part which is the slowest to help.
@@ -368,7 +414,7 @@ class AutoDownload:
         while True:
             try:
                 if part.stream==None:
-                    part.stream=requests.get(self.url,headers=header,stream=True)
+                    part.stream=requests.get(self.url,headers=header,stream=True,timeout=self.timeout)
                 if part.stream.status_code//100 not in [2,3]:
                     raise ConnectError(self.url)
                 with open(part.fileName,"wb") as f:
@@ -404,9 +450,9 @@ class AutoDownload:
                     self._finished()
                     return
             except BaseException as err:
+                if retryNum==self.maxThreadRetry:
+                    self._errorShower(err)
                 retryNum+=1
-                if retryNum>=self.maxRetry:
-                    raise err
-                self.statue=f"retry {retryNum}"
+                part.statue=f"retry {retryNum}"
                 part.statueNum=1
                 self._logShower("Part %d %s:%s"%(partNum,err.__class__.__name__,str(err)),level=logging.WARNING)
